requirements.txt
# runtime

numpy
pillow
rawkit
opencv-contrib-python
imreg_dft



# development
matplotlib
pytest
ipython